# src/api/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "Proyecto de Backend para pagina E-Commerce ",
      "version": "0.1.0",
      "title": "EcommerceBackend"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "schemes": ["http"],
...
    "paths": {
      "/products": {
        "post": {
          "tags": ["products"],
          "summary": "add product",
          "description": "add product",
          "operationId": "addProduct",
          "produces": ["application/json"],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "product object that needs to be added",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Product_variant1"
              }
            }
          ],
          "responses": {
            "default": {
              "description": "successful operation"
            }
          }
        },
        "get": {
          "tags": ["products"],
          "summary": "get all products",
          "description": "get all products",
          "operationId": "getAllProducts",
          "responses": {
            "default": {
              "description": "successful operation"
            }
          }
        }
      },
      "/products/{id}": {
        "get": {
          "tags": ["products"],
          "summary": "get a product by id",
          "description": "get a product by id",
          "operationId": "getProductById",
          "responses": {
            "default": {
              "description": "successful operation"
            }
          }
        },
        "put": {
          "tags": ["products"],
          "summary": "update a product by id",
          "description": "update a product by id",
          "operationId": "updateProduct",
          "produces": ["application/json"],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "product object that needs to be updated",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Product_variant1"
              }
            }
          ],
          "responses": {
            "default": {
              "description": "successful operation"
            }
          }
        },
        "delete": {
          "tags": ["products"],
          "summary": "delete a product by id",
          "description": "delete a product by id",
          "operationId": "deleteProduct",
          "produces": ["application/json"],
          "responses": {
            "default": {
              "description": "successful operation"
            }
          }
        }
      },
...
    },
    "definitions": {
...
      "Product_variant1": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Computadora"
          },
          "description": {
            "type": "string",
            "example": "Notebook HP 8GB RAM"
          },
          "code": {
            "type": "string",
            "example": "405fth"
          },
          "thumbnail": {
            "type": "string",
            "example": "https://res.cloudinary.com/dnwlad3kp/image/upload/v1637962406/compragamer_Imganen_general_21800_Notebook_ASUS_x509_15_udyq3y.png"
          },
          "stock": {
            "type": "number",
            "example": 20
          },
          "price": {
            "type": "number",
            "example": 1000
          }
        }
      }
    }
}